+++ src/pprune/publication_maps.py
@@ -214,9 +214,9 @@ class AirIndia171(PublicationMap):
 
     def get_set_of_removed_subjects(self) -> typing.Set[str]:
         return {
-            "RAT (Alternate Noise Sources)",
             '51 Day Issue',
             'Audio Analysis',
+            'Biocide',
             'Bird Strike',
             'Engine Over-speed (All)',
             'Engine Shutdown (Over-speed)',
@@ -236,6 +236,7 @@ class AirIndia171(PublicationMap):
             'MLG (All)',
             'MLG Tilt',
             'N2 Over-speed',
+            'RAT (Alternate Noise Sources)',
             'RAT (Sound)',
             'RAT (Witnesses)',
             'Suicide/Murder',
@@ -291,6 +292,9 @@ class AirIndia171(PublicationMap):
         'murder': 'Suicide/Murder',
         'jumpseaters': 'Jump Seat',
         'jumpseat': 'Jump Seat',
+        'moderators': 'Thread Moderation',
+        'moderation': 'Thread Moderation',
+        'mods': 'Thread Moderation',
     }
     # This maps capitilised words (stripped of punctuation) to their subject.
     # Any post that has that capitilised word in it is treated as part of that subject.
@@ -634,4 +638,6 @@ class AirIndia171(PublicationMap):
         ('tdracer on Air Accident Investigations', 11924194,),
         # https://www.pprune.org/accidents-close-calls/667141-preliminary-air-india-crash-report-published-65.html#post11924281
         ('787 Maintenance and those Fuel Cutoff Switches (15:00 onwards)', 11924281,),
+        ('NTSB chair Jennifer Homendy on the Preliminary Report', 11925921,),
+        ('The International Federation of Air Line Pilots\' Associations on the Preliminary Report', 11925980,),
     )